Refactor: Integrate MainLayout from layouts folder into router with Signup route
