+++ client/src/layouts/MainLayout.jsx
@@ -1,0 +1,15 @@
+import { Outlet } from "react-router-dom";
+import Navbar from "../components/Navbar";
+
+const MainLayout = () => {
+	return (
+		<>
+			<Navbar />
+			<main className="p-4">
+				<Outlet />
+			</main>
+		</>
+	);
+};
+
+export default MainLayout;
