+++ client/src/router.jsx
@@ -1,10 +1,17 @@
 import { createBrowserRouter } from "react-router-dom";
-import Signup from "./components/Signup";
+import MainLayout from "./layouts/MainLayout";
+import Signup from "./pages/Signup";
 
 const router = createBrowserRouter([
 	{
-		path: "/signup",
-		element: <Signup />,
+		path: "/",
+		element: <MainLayout />,
+		children: [
+			{
+				path: "signup",
+				element: <Signup />,
+			},
+		],
 	},
 ]);
